Moved files to the new Modules folder to contain classes not directly related to HES
Current folder structure required for hes to operate, will throw exceptions if its wrong (Folders.cs will be used to make sure the folders are structured this way soon.

    /// ServerFolder/
    ///     ...HellionExtendedServer.exe
    ///     ...HellionExtendedServer.exe.config
    ///     ...Hes/ (Using this folder helps keep server hosters server directory cleaned up (2 files, 1 folder))
    ///         ...Config.cfg (HES config file)
    ///         ...localization (Contains resources for languages)
    ///         ...logs (contains HES's logs)
    ///         ...updates (will be used as a temp directory for updates)
    ///         ...bin/ (any files that HES or HES.Common references to keep the server dir clean)
    ///             ...HellionExtendedServer.Common.dll
    ///             ...NLog.dll
    ///             ...NLog.dll.config
