+++ HellionExtendedServer/HellionExtendedServer.csproj
@@ -126,8 +126,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Config.cs" />
-    <Compile Include="Localization.cs" />
+    <Compile Include="Modules\Config.cs" />
+    <Compile Include="Modules\Folders.cs" />
+    <Compile Include="Modules\Localization.cs" />
     <Compile Include="Managers\CommandManager.cs" />
     <Compile Include="Managers\Plugins\PluginInfo.cs" />
     <Compile Include="Managers\ServerInstance.cs" />
@@ -145,7 +146,7 @@
     <Compile Include="Managers\Plugins\PluginManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Managers\UpdateManager.cs" />
+    <Compile Include="Modules\Updater.cs" />
     <EmbeddedResource Include="HESGui.resx">
       <DependentUpon>HESGui.cs</DependentUpon>
     </EmbeddedResource>
