+++ HellionExtendedServer/Modules/Folders.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HellionExtendedServer.Modules
+{
+    /// <summary>
+    /// This class will be used to make sure the folders for hes exist and have hes's required libraries contained within.
+    /// This needs to be completed before we can release it, incase people delete the HES folder or the update fails, etc!!
+    /// 
+    /// HellionDedicatedServer/
+    ///     HellionExtendedServer.exe
+    ///     HellionExtendedServer.exe.config    
+    ///     Hes/
+    ///         Config.cfg
+    ///         localization
+    ///         logs
+    ///         updates
+    ///         bin/
+    ///             HellionExtendedServer.Common.dll
+    ///             NLog.dll
+    ///             NLog.dll.config
+    /// </summary>
+    public class Folders
+    {
+        public Folders()
+        {
+
+        }
+
+    }
+}
